orchestration/src/retriever.py
"""
Hybrid retrieval: vector (Milvus) + graph (Neo4j) + SQL (PostgreSQL).
"""
from typing import List, Dict, Any
from pymilvus import Collection
from neo4j import GraphDatabase
import psycopg2


class HybridRetriever:
    def __init__(self, pg_conn_str: str, milvus_host: str, milvus_port: int, neo4j_uri: str, neo4j_user: str, neo4j_password: str):
        self.pg_conn_str = pg_conn_str
        self.milvus_host = milvus_host
        self.milvus_port = milvus_port
        self.neo4j_driver = GraphDatabase.driver(neo4j_uri, auth=(neo4j_user, neo4j_password))
    
    def vector_search(self, embedding: List[float], top_k: int = 5) -> List[Dict[str, Any]]:
        """Search Milvus for similar documents."""
        # Placeholder: Initialize Milvus collection and search
        # In production, use pymilvus Collection to query
        return []
    
    # def graph_search(self, account_id: str, depth: int = 2) -> List[Dict[str, Any]]:
    #     """Traverse Neo4j for related accounts, people, signals."""
    #     with self.neo4j_driver.session() as session:
    #         query = f"""
    #             MATCH (acc:Account {{id: $acc_id}})
    #             CALL apoc.path.expandConfig(acc, {{
    #                 relationshipFilter: '>moves|>signals',
    #                 maxLevel: {depth}
    #             }}) YIELD path
    #             RETURN path
    #             LIMIT 10
    #         """
    #         result = session.run(query, acc_id=account_id)
    #         paths = [record["path"] for record in result]
    #         return self._extract_path_context(paths)
    
    def graph_search(self, account_id: str, depth: int = 2):
        """
        Retrieve relationship map for account.
        MOVED TO STUB: Neo4j connection disabled.
        """
        logger.info(f"Graph search requested for {account_id} (Neo4j Bypassed)")
        
        # Return mock relationships so the UI still looks good
        return [
            {"from": account_id, "to": "Key Stakeholder", "relation": "decision_maker"},
            {"from": account_id, "to": "CompetitorX", "relation": "competes_with"},
        ]
    
    def sql_search(self, account_id: str, limit: int = 10) -> List[Dict[str, Any]]:
        """Query PostgreSQL for structured metadata."""
        conn = psycopg2.connect(self.pg_conn_str)
        cur = conn.cursor()
        cur.execute(
            "SELECT id, title, source, url FROM documents WHERE account_id = %s ORDER BY created_at DESC LIMIT %s",
            (account_id, limit)
        )
        rows = cur.fetchall()
        cur.close()
        conn.close()
        
        return [
            {"id": r[0], "title": r[1], "source": r[2], "url": r[3]}
            for r in rows
        ]
    
    # def hybrid_search(self, embedding: List[float], account_id: str, query: str) -> Dict[str, List]:
    #     """Combine vector, graph, and SQL search."""
    #     return {
    #         "vector_results": self.vector_search(embedding, top_k=5),
    #         "graph_context": self.graph_search(account_id, depth=2),
    #         "sql_metadata": self.sql_search(account_id, limit=10),
    #     }
    
    def hybrid_search(self, embedding, account_id, query):
        """
        Updated to skip Neo4j during hybrid retrieval
        """
        # ... your existing vector/SQL code ...
        
        return {
            "vector_results": [], # or your actual vector results
            "sql_metadata": {},   # or your actual SQL results
            "graph_context": []   # Set this to empty to avoid calling Neo4j
        }
    
    def _extract_path_context(self, paths: List) -> List[Dict[str, Any]]:
        """Convert Neo4j paths to readable context."""
        # Placeholder implementation
        return []
    
    def close(self):
        if self.neo4j_driver:
            self.neo4j_driver.close()
