orchestration/pyproject.toml
[tool.poetry]
name = "salesai-orchestration"
version = "0.1.0"
description = "LangChain + GROQ orchestrator for hybrid RAG"
authors = ["SalesAI Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
groq = "^1.0.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
python-dotenv = "^1.0"
pika = "^1.3"
requests = "^2.31"
psycopg2-binary = "^2.9"
pymilvus = "^2.3"
neo4j = "^5.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
black = "^23.0"
mypy = "^1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
